Creation and operation of the MySQL database.
Create:
- SignalDataBase class for creating and dealing with the MySQL database.
This database contains the application signals and their names (long
texts of the kuka robot) for all technologies specified by the customer.
- script in the Editor class to automatically connect to a database or
to create a new database if it does not exist.
- "Create new database" action in the "File" menu and its functionality.
This action allows to create a new database from the specified CSV files
included in the corporate standard.
- ProgressDialog class to show a progress of creating a new database.
- credentials.py file which contains the data needed to connect to the
database.